authentication corrected

+++ app/Http/Controllers/PostsController.php
@@ -19,6 +19,15 @@ class PostsController extends Controller
     // delete 
     //php artisan make:controller PostsController --resource
 
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
 
 
     /**
@@ -96,6 +105,10 @@ class PostsController extends Controller
     {
         $post = Post::find($id);
 
+        if (auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized page');
+        }
+
         return view('posts.edit')->with('post',$post);
 
     }
@@ -130,6 +143,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if (auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized page');
+        }
+        
         $post->delete();
         return redirect('/posts')->with('success','Post Deleted');
     }
